requirements_test.txt
#### TESTING-RELATED PACKAGES ####

# Checks style, syntax, and other useful errors.
# We need this commit because of https://github.com/PyCQA/pylint/pull/1430,
# unpin when pylint 1.7.3 is released.
git+git://github.com/PyCQA/pylint.git@2561f539d60a3563d6507e7a22e226fb10b58210

# We'll use pytest to run our tests; this isn't really necessary to run the code, but it is to run
# the tests.  With this here, you can run the tests with `py.test` from the base directory.
pytest

# Static type checking
mypy

# Makes it so that pytest can handle the code structure we use, with src/main/python, and src/test.
pytest-pythonpath

# Allows generation of coverage reports with pytest.
pytest-cov

# Allows marking tests as flaky, to be rerun if they fail
flaky

# Allows codecov to generate coverage reports
coverage
codecov

# Required to run sanic tests
aiohttp

# Required to mock out `requests` calls
responses>=0.7

#### DOC-RELATED PACKAGES ####

# Builds our documentation.
sphinx==1.5.3

# Watches the documentation directory and rebuilds on changes.
sphinx-autobuild

# enables numpydoc formatting in sphinx
numpydoc

# doc theme
sphinx_rtd_theme

# Only used to convert our readme to reStructuredText on Pypi.
pypandoc

# Pypi uploads
twine
